backend/foodgram/requirements.txt
asgiref==3.5.2
Django==4.1
django-filter==22.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
flake8==5.0.4
isort==5.10.1
mccabe==0.7.0
Pillow==9.2.0
psycopg2-binary==2.8.6
pycodestyle==2.9.1
pyflakes==2.5.0
PyJWT==2.4.0
python-dotenv==0.20.0
pytz==2022.1
sqlparse==0.4.2
tzdata==2022.1
